Use Review Board's special extension packaging setup method.
We were having difficulty getting this extension packaged, and
it looks like this is because we need to use this special "setup"
method that does the job of pre-compiling the LESS code and doing
other static file prep.

It also flips off zip_safe, which is likely another problem point
for our installation.